crlf_data: add corpus of known-good odb-filtered data
Use the crlf data scripts to produce a corpus of known-good data in
"git" format (aka ODB format) from a variety of files with different
line endings.  `git` created these files running `git add` to stage the
contents then extracting the data from the repository.

We'll use these to ensure that we create identical contents when we add
files into the index.