smart subtransport: don't free the url when resetting the stream
We reset the stream in the RPC (HTTP) case constantly and then pass the url
field to the action method for the transport to perform their action.

A transport would thus get `NULL` for the url and if it does use it, it might
crash or at least return an error making it impossible for it to work.

The transports that come with libgit2 keep their own copy of the url and thus
this issue was not noticed.