refdb_fs: load packed-refs only when necessary
When checking if a ref exists, we load the packed-refs file and then
return true if the ref is found either loose or in the packed-refs
cache.

This is subject to the same race described in the previous commit, where
a simultaneous pack-refs may migrate a ref from loose to packed and
cause us to see neither. We can fix it by making sure we load
packed-refs only after checking the loose store.

As a bonus, this is a minor optimization: we'll avoid even looking at
the packed-refs file if we see a loose version.