regex: allow regex selection in cmake
Users can now select which regex implementation they want to use: one of
the system `regcomp_l`, the system PCRE, the builtin PCRE or the
system's `regcomp`.

By default the system `regcomp_l` will be used if it exists, otherwise
the system PCRE will be used.  If neither of those exist, then the
builtin PCRE implementation will be used.

The system's `regcomp` is not used by default due to problems with
locales.